Readme Generator

+++ scripts/2_install/1a_install_mariadb_centos.sh
@@ -27,9 +27,14 @@ gpgcheck=1" > /etc/yum.repos.d/mariadb_${VERSION}.repo
 
 cmd "cat /etc/yum.repos.d/mariadb_${VERSION}.repo"
 
-cmd "yum -y remove mysql-server mariadb-server"
-lRC=$(($lRC + $?))
+if [ "$force" = "1" ]; then
+	cmd "yum -y remove mysql-server mariadb-server"
+	lRC=$(($lRC + $?))
+fi
 
+cmd "mkdir  -p /var/log/mysql /run/mysqld"
+cmd "chown -R mysql.mysql /var/log/mysql /run/mysqld"
+lRC=$(($lRC + $?))
 
 cmd "yum -y install https://dl.fedoraproject.org/pub/epel/epel-release-latest-${VERSION_ID}.noarch.rpm"
 lRC=$(($lRC + $?))
@@ -40,21 +45,24 @@ lRC=$(($lRC + $?))
 cmd "yum -y install cracklib cracklib-dicts tree socat jemalloc rsync nmap lsof perl-DBI nc mariadb-server-utils pigz perl-DBD-MySQL git pwgen"
 lRC=$(($lRC + $?))
 
-cmd "yum -y install https://repo.percona.com/yum/release/${VERSION_ID}/RPMS/x86_64/percona-toolkit-3.2.1-1.el8.x86_64.rpm"
+cmd "yum -y install https://repo.percona.com/yum/release/latest/RPMS/x86_64/percona-toolkit-3.2.1-1.el6.x86_64.rpm"
 lRC=$(($lRC + $?))
 
 cmd "pip3 install mycli"
 lRC=$(($lRC + $?))
 
-cmd "yum -y install https://rpmfind.net/linux/fedora-secondary/development/rawhide/Everything/s390x/os/Packages/m/mysqlreport-3.5-23.fc33.noarch.rpm"
+cmd "yum -y install https://rpmfind.net/linux/fedora-secondary/development/rawhide/Everything/s390x/os/Packages/m/mysqlreport-3.5-24.fc34.noarch.rpm"
 lRC=$(($lRC + $?))
 
 [ -d "/opt/local" ] || cmd "mkdir -p /opt/local"
 
-[ -d "/opt/local/MySQLTuner-perl" ] && cmd "rm -rf /opt/local/MySQLTuner-perl"
-
-cd /opt/local
-cmd "git clone https://github.com/major/MySQLTuner-perl.git"
+if [ -d "/opt/local/MySQLTuner-perl" ]; then
+	cd /opt/local/MySQLTuner-perl
+	cmd "git pull"
+else
+	cd /opt/local
+	cmd "git clone https://github.com/major/MySQLTuner-perl.git"
+fi
 lRC=$(($lRC + $?))
 
 cmd "chmod 755 /opt/local/MySQLTuner-perl/mysqltuner.pl"
@@ -63,6 +71,6 @@ echo 'export PATH=$PATH:/opt/local/MySQLTuner-perl' > /etc/profile.d/mysqltuner.
 chmod 755 /etc/profile.d/mysqltuner.sh
 
 cmd "yum -y install perl-App-cpanminus"
-cmd "cpanm MySQL::Diff"
+#cmd "cpanm MySQL::Diff"
 footer "END SCRIPT: $NAME"
 exit $lRC
